[MCS251] Land T1 test suite: runner + template + 7 pilots + 15 kernels + probes (Shizuku/Momo/Sakuna)
The demo x QEMU test system's T1 (algorithm-kernel regression) tier is now
complete and verified on the 67a17057 baseline (post-endianness-fix).

run-tests.py (Shizuku): --filter t1|t2|t4|all, dual-oracle + DUT triangle,
frozen-binary pinned to versioned path, JSON results, two retireable shims
(--no-ir-shims: IR-name adaptation @name->_name; const-shift lowering with a
volatile-alloca barrier) flagged with retirement conditions (the shims are
already obsolete for the fixed widths and will go once the real shift
lowering/P-C naming land).

t1/TEMPLATE.md: kernel.c width is a hard rule (u16 = unsigned short; a bare
unsigned int diverges host int=32 vs SDCC int=16), known compiler-edge
avoidances, wrapper/host-main/EXTRACT standard parts + run.py's
mangle_ll_symbols() / shift shims, the '.L'/CONST=0xFC8000 traps, smoke =
no host-main.

7 pilots instantiated (kernel.c+wrapper.c+host-main.c+gen-vectors.py):
const-handle, judge-type, string-length, sum-sfn, led8-nibble (faithful
u16), nec-decode (ISR state machine, 2-frame vectors), delay-smoke. All
7 PASS on the 67a17057 baseline (0 FAIL); leading serials per case in
RESULTS.md.

15 extracted kernels (Momo): adc-key, decimal/hex-parse, keyboard-scan,
operator-precedence, pulse-width, reverse16/32, rtc-step, stack-packed/
pop-packed, task-dispatch/marks, temperature-lookup, wildcard-recursive -
all kernel.c+EXTRACT.md, pending-instantiation (no wrapper/host yet).

probes/ (Sakuna): endianness / endianness-fail / combiner-fold / nec-frame2
+ per-dir README (tool md5s, phenomenon, expected-serial correction - the
endianness-fail true expectation is Bl34h12) + current-baseline fixed-
load-i16 artifacts; no /tmp dependency; probe archive manifest + sha256.

DESIGN.md gains the implementation-record appendix (7. section).

+++ validation/mcs251-demo-test/t1/operator-precedence/kernel.c
@@ -4,7 +4,7 @@ typedef unsigned short u16;
 #if defined(__SDCC_mcs251)
 typedef unsigned long u32;
 #else
-typedef unsigned int u32;
+typedef unsigned long u32;
 #endif
 #define LEVEL_BIGGER 0u
 #define LEVEL_SMALLER 1u
